+++ src/mscore/mbrola.hpp
@@ -53,7 +53,7 @@ private:
   bool check_voice(const std::string& voice_name);
 
   // Make up voice parameters for backend:
-  void voicify(double rate, double pitch = 1.0);
+  void voicify(float rate, float pitch = 1.0);
 };
 
 } // namespace multispeech
